+++ app/Container/Financial/src/Controllers/Management/ManageStatusController.php
@@ -5,6 +5,7 @@ namespace App\Container\Financial\src\Controllers\Management;
 
 use App\Container\Financial\src\Repository\StatusRequestRepository;
 use App\Container\Financial\src\Requests\StatusRequest\StoreStatusRequestsRequest;
+use App\Container\Overall\Src\Facades\AjaxResponse;
 use App\Http\Controllers\Controller;
 
 class ManageStatusController extends Controller
@@ -30,7 +31,10 @@ class ManageStatusController extends Controller
      */
     public function index()
     {
-        return view('financial.management.status.index');
+        if ( request()->isMethod('GET') )
+            return view('financial.management.status.index');
+
+        return abort( 405 );
     }
 
     /**
@@ -41,9 +45,12 @@ class ManageStatusController extends Controller
      */
     public function store(StoreStatusRequestsRequest $request)
     {
-        return ( $this->statusRequestRepository->store( $request ) ) ?
-                jsonResponse() :
-                jsonResponse('error', 'processed_fail', 422);
+        if ( request()->isMethod('POST') )
+            return ( $this->statusRequestRepository->store( $request ) ) ?
+                    jsonResponse() :
+                    jsonResponse('error', 'processed_fail', 422);
+
+        return AjaxResponse::make(__('javascript.http_status.error', ['status' => 405]), __('javascript.http_status.method', ['method' => 'POST']), '', 405);
     }
 
     /**
@@ -55,13 +62,16 @@ class ManageStatusController extends Controller
      */
     public function update(StoreStatusRequestsRequest $request, $id)
     {
-        $status = $this->statusRequestRepository->getModel()->find( $id );
-        if ( isset( $status->{ status_name() } ) && isEditableStatus( $status->{ status_name() }  ) ) {
-            return ( $this->statusRequestRepository->update( $request, $id ) ) ?
-                jsonResponse('success', 'updated_done', 200) :
-                jsonResponse('error', 'updated_fail', 422);
+        if ( request()->isMethod('PUT') || request()->isMethod('PATCH') ) {
+            $status = $this->statusRequestRepository->getModel()->find($id);
+            if (isset($status->{status_name()}) && isEditableStatus($status->{status_name()})) {
+                return ($this->statusRequestRepository->update($request, $id)) ?
+                    jsonResponse('success', 'updated_done', 200) :
+                    jsonResponse('error', 'updated_fail', 422);
+            }
+            return jsonResponse('error', 'updated_fail_status', 422);
         }
-        return jsonResponse('error', 'updated_fail_status', 422);
+        return AjaxResponse::make(__('javascript.http_status.error', ['status' => 405]), __('javascript.http_status.method', ['method' => 'PUT / PATCH']), '', 405);
     }
 
     /**
@@ -72,13 +82,18 @@ class ManageStatusController extends Controller
      */
     public function destroy($id)
     {
-        $status = $this->statusRequestRepository->getModel()->find( $id );
-        if ( isset( $status->{ status_name() } ) && isEditableStatus( $status->{ status_name() }  ) ) {
-            return ($this->statusRequestRepository->destroy($id)) ?
-                jsonResponse('success', 'deleted_done', 200) :
-                jsonResponse('error', 'deleted_fail', 422);
+        if ( request()->isMethod('DELETE') ) {
+            $status = $this->statusRequestRepository->getModel()->find($id);
+            if (isset($status->{status_name()}) && isEditableStatus($status->{status_name()})) {
+                return ($this->statusRequestRepository->destroy($id)) ?
+                    jsonResponse('success', 'deleted_done', 200) :
+                    jsonResponse('error', 'deleted_fail', 422);
+            }
+
+            return jsonResponse('error', 'destroy_fail_status', 422);
+
         }
 
-        return jsonResponse('error', 'destroy_fail_status', 422);
+        return AjaxResponse::make(__('javascript.http_status.error', ['status' => 405]), __('javascript.http_status.method', ['method' => 'DELETE']), '', 405);
     }
 }
